refactor games module
